rename package from zio.http.html to zio.http.template to avoid ambiguity
