+++ zio-http-example/src/main/scala/example/StaticServer.scala
@@ -4,7 +4,7 @@ import zio.ZIOAppDefault
 
 import zio.http._
 import zio.http.codec.PathCodec.trailing
-import zio.http.html._
+import zio.http.template._
 
 object StaticServer extends ZIOAppDefault {
