+++ zio-http/src/main/scala/zio/http/Handler.scala
@@ -17,7 +17,7 @@
 package zio.http
 
 import zio._
-import zio.http.html.{Html, Template}
+import zio.http.template.{Html, Template}
 import zio.http.Header.HeaderType
 import zio.http.internal.HeaderModifier
 import zio.stream.ZStream
