+++ zio-http/src/main/scala/zio/http/template/package.scala
@@ -14,32 +14,6 @@
  * limitations under the License.
  */
 
-package zio.http.html
+package zio.http
 
-/**
- * A ZIO Http styled general purpose templates
- */
-object Template {
-
-  def container(heading: CharSequence)(element: Html): Html = {
-    html(
-      head(
-        title(s"ZIO Http - ${heading}"),
-        style("""
-                | body {
-                |   font-family: monospace;
-                |   font-size: 16px;
-                |   background-color: #edede0;
-                | }
-                |""".stripMargin),
-      ),
-      body(
-        div(
-          styles := Seq("margin" -> "auto", "padding" -> "2em 4em", "max-width" -> "80%"),
-          h1(heading),
-          element,
-        ),
-      ),
-    )
-  }
-}
+package object template extends Attributes with Elements {}
